+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -543,10 +543,21 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "lootIncludeClueScrolls",
+        name = "Include Clue Loot",
+        description = "Allow notifications for loot from Clue Scrolls",
+        position = 36,
+        section = lootSection
+    )
+    default boolean lootIncludeClueScrolls() {
+        return true;
+    }
+
+    @ConfigItem(
         keyName = "lootNotifMessage",
         name = "Notification Message",
         description = "The message to be sent through the webhook. Use %USERNAME% to insert your username, %LOOT% to insert the loot and %SOURCE% to show the source of the loot",
-        position = 36,
+        position = 37,
         section = lootSection
     )
     default String lootNotifyMessage() {